+++ mixer/bl_panels.py
@@ -1,3 +1,7 @@
+"""
+This module define Blender Panels and UI types for the addon.
+"""
+
 from __future__ import annotations
 from typing import TYPE_CHECKING