+++ mixer/bl_utils.py
@@ -1,3 +1,7 @@
+"""
+This modules defines utilities related to Blender.
+"""
+
 from __future__ import annotations
 from typing import TYPE_CHECKING
 import bpy