+++ mixer/blender_client/__init__.py
@@ -1,3 +1,21 @@
+"""
+This package defines how we send Blender updates to the server, and how we interpret updates we receive to update
+Blender's data.
+
+This functionnalities are implemented in the BlenderClient class and in submodules of the package.
+
+Submodules with a well defined entity name (camera, collection, light, ...) handle updates for the corresponding
+data type in Blender. The goal is to replace all this specific code with the submodule data.py, which use
+the blender_data package to treat updates of Blender's data in a generic way.
+
+Specific code will still be required to handle non-Blender clients. As an exemple, mesh.py add to the MESH
+message a triangulated, with modifiers applied, of the mesh. This is for non-Blender clients. In the future we want to
+move these kind of specific processes to a plug-in system.
+
+The main function in this file is BlenderClient.network_consumer, which is the function called by the Blender timer
+we register.
+"""
+
 import logging
 import os
 import struct
@@ -20,6 +38,7 @@ from mixer.blender_client import material as material_api
 from mixer.blender_client import mesh as mesh_api
 from mixer.blender_client import object_ as object_api
 from mixer.blender_client import scene as scene_api
+from mixer.log_utils import log_traceback
 import mixer.shot_manager as shot_manager
 from mixer.stats import stats_timer
 from mixer.draw_handlers import set_draw_handlers
@@ -64,6 +83,12 @@ class SendSceneContentFailed(Exception):
 
 
 class BlenderClient(Client):
+    """
+    Client specialized for Blender. Extends Client base class by adding and handling data related to Blender.
+
+    See network_consumer() method which can be considered the entry point of this class.
+    """
+
     def __init__(self, host=common.DEFAULT_HOST, port=common.DEFAULT_PORT):
         super(BlenderClient, self).__init__(host, port)
 
@@ -607,10 +632,25 @@ class BlenderClient(Client):
                                 areas_3d.append(view_id)
                 windows.append({"scene": scene, "view_layer": view_layer, "screen": screen, "areas_3d": areas_3d})
 
+        # Documentation to update if you change "blender_windows": doc/protocol.md
         return {"blender_windows": windows, common.ClientAttributes.USERSCENES: scene_attributes}
 
     @stats_timer(share_data)
     def network_consumer(self):
+        """
+        This method can be considered the entry point of this class. It is meant to be called regularly to send
+        pending commands to the server, and receive then process new ones.
+
+        Pending commands are accumulated with add_command(), most calls originate from handlers function.
+
+        Incoming commands are read from the socket and directly processed here to update Blender's data. This can
+        be costly and a possible optimization in the future would be to split the processing accross several timer
+        run. This can be challenging because we need to keep the current update state. Maybe this can be solved
+        naturally with coroutines.
+
+        We call it from the timer registered by the addon.
+        """
+
         from mixer.bl_panels import redraw as redraw_panels, update_ui_lists
 
         assert self.is_connected()
@@ -679,6 +719,7 @@ class BlenderClient(Client):
                             assert share_data.client.current_room is not None
                             self.set_room_attributes(
                                 share_data.client.current_room,
+                                # Documentation to update if you change "experimental_sync": doc/protocol.md
                                 {"experimental_sync": get_mixer_prefs().experimental_sync},
                             )
                             send_scene_content()
@@ -803,7 +844,8 @@ class BlenderClient(Client):
                         self.skip_next_depsgraph_update = True
 
                 except Exception as e:
-                    logger.warning(f"Exception during processing of message {str(command.type)} ...\n", stack_info=True)
+                    logger.warning(f"Exception during processing of message {str(command.type)}")
+                    log_traceback(logger.warning)
                     if get_mixer_prefs().env == "development" or isinstance(e, SendSceneContentFailed):
                         raise
 
@@ -881,6 +923,10 @@ def update_params(obj):
 
 
 def clear_scene_content():
+    """
+    Clear data before joining a room.
+    """
+
     from mixer.handlers import HandlerManager
 
     with HandlerManager(False):
@@ -919,6 +965,10 @@ def clear_scene_content():
 
 @stats_timer(share_data)
 def send_scene_content():
+    """
+    Initial data send to fill a new room.
+    """
+
     from mixer.handlers import HandlerManager, send_scene_data_to_server
 
     if get_mixer_prefs().no_send_scene_content: