+++ mixer/blender_data/specifics.py
@@ -24,26 +24,37 @@ def bpy_data_ctor(collection_name: str, proxy: BpyIDProxy) -> Union[T.ID, None]:
         image = None
         if is_packed:
             name = proxy.data("name")
-            size = proxy.data("size")
-            width = size.data(0)
-            height = size.data(1)
+            width, height = proxy.data("size")
             image = collection.new(name, width, height)
             # remaning attributes will be saved from the received proxy attributes
         else:
             path = proxy.data("filepath")
             if path != "":
-                image = collection.load(path)
+                try:
+                    image = collection.load(path)
+                except RuntimeError as e:
+                    logger.warning(f'Cannot load image at path "{path}". Exception: ')
+                    logger.warning(f"... {e}")
+                    return None
                 # we may have received an ID named xxx.001 although filepath is xxx, so fix it now
                 image.name = proxy.data("name")
         return image
 
     if collection_name == "objects":
         name = proxy.data("name")
+        target = None
         target_proxy = proxy.data("data")
         if target_proxy is not None:
-            target = target_proxy.target()
-        else:
-            target = None
+            # use class name to work around circular references with proxy.py
+            target_proxy_class = target_proxy.__class__.__name__
+            if target_proxy_class != "BpyIDRefProxy":
+                # error on the sender side
+                logger.warning(
+                    f"bpy.data.objects[{name}].data proxy is a {target_proxy_class}. Expected a BpyIDRefProxy"
+                )
+                logger.warning("... loaded as Empty")
+            else:
+                target = target_proxy.target()
         object_ = collection.new(name, target)
         return object_