+++ mixer/stats.py
@@ -1,3 +1,7 @@
+"""
+This module defines utilities for statistics computation and export.
+"""
+
 import time
 import os
 import json
@@ -7,10 +11,16 @@ import tempfile
 from pathlib import Path
 import functools
 
+from mixer.os_utils import getuser
+
 logger = logging.getLogger(__name__)
 
 
 class StatsTimer:
+    """
+    Class to measure execution time of a code section. Can be used as a context manager.
+    """
+
     def __init__(self, share_data, key, log=None):
         assert share_data.current_statistics
 
@@ -67,7 +77,7 @@ class StatsTimer:
 
 def get_stats_directory():
     if "MIXER_USER_STATS_DIR" in os.environ:
-        username = os.getlogin()
+        username = getuser()
         base_shared_path = Path(os.environ["MIXER_USER_STATS_DIR"])
         if os.path.exists(base_shared_path):
             return os.path.join(os.fspath(base_shared_path), username)
@@ -107,6 +117,11 @@ def save_statistics(stats_dict, stats_directory):
 
 
 def stats_timer(share_data, log=None):
+    """
+    A decorator to measure and store execution time of a function with respec to
+    the call stack.
+    """
+
     def inner_decorator(func):
         @functools.wraps(func)
         def wrapper(*args, **kwargs):